Chat: I wanna create a simple to serve as a introduction... (automagik-forge 46d63d6c)
I wanna create a simple to serve as a introduction to this github project: https://github.com/amorabot/Inscripted. As a guideline, you can use these 2 websites: https://spellforged.github.io/ and https://face.land/. The project provided is a RPG server for Minecraft and i want the website to briefly showcase what the server is about, with sections for: a introduction, the item system, for the archetype system and concepts (segmented per-archetype and their colors as accents) and, lastly, who is the developer (me), including a way to join the discord (https://discord.gg/SWnWghYRHr). There will be 2 main images attached. One of them is the main "banner" for the project with a black & white theme with orange accents. Make that the main theme for the website, trying to make it dark-themed but avoid using absolute black. The other one is a pixel art of me, use that to illustrate the developer wherever needed. Also, to fit the "gothic" theme, use the "Alkhemical" font files in the repository.